refactor(use data value mutation): split into separate hooks
