+++ src/data-workspace/inputs/boolean-radios.js
@@ -3,7 +3,7 @@ import { Button, Radio } from '@dhis2/ui'
 import cx from 'classnames'
 import React, { useState } from 'react'
 import { useField, useForm } from 'react-final-form'
-import { useDataValueMutation } from '../data-entry-cell/use-data-value-mutation.js'
+import { useSetDataValueMutation } from '../use-data-value-mutation/index.js'
 import styles from './inputs.module.css'
 import { convertCallbackSignatures, InputPropTypes } from './utils.js'
 
@@ -54,7 +54,7 @@ export const BooleanRadios = ({
     const form = useForm()
 
     const [lastSyncedValue, setLastSyncedValue] = useState()
-    const { mutate } = useDataValueMutation()
+    const { mutate } = useSetDataValueMutation()
     const syncData = (value) => {
         // todo: Here's where an error state could be set: ('onError')
         mutate(
