+++ src/data-workspace/use-data-value-mutation/add-data-value.js
@@ -1,0 +1,10 @@
+// Adds dataValue without mutating previousDataValueSet
+export default function addDataValue(previousDataValueSet, newDataValue) {
+    return {
+        ...previousDataValueSet,
+        // dataValueSet.dataValues can be undefined
+        dataValues: previousDataValueSet.dataValues
+            ? [...previousDataValueSet.dataValues, newDataValue]
+            : [newDataValue],
+    }
+}
