+++ src/data-workspace/use-data-value-mutation/update-data-value.js
@@ -1,0 +1,14 @@
+// Updates dataValue without mutating previousDataValueSet
+export default function updateDataValue(
+    previousDataValueSet,
+    updatedDataValue,
+    targetIndex
+) {
+    const newDataValues = [...previousDataValueSet.dataValues]
+    newDataValues[targetIndex] = updatedDataValue
+
+    return {
+        ...previousDataValueSet,
+        dataValues: newDataValues,
+    }
+}
